Separate Tabular Preprocessing
